+++ src/common/persistence/OperationQueue.ts
@@ -27,35 +27,46 @@ export default class OperationQueue implements IOperationQueue {
    *
    * @template T
    * @param {Promise<T>} op
+   * @param {string} [contextId]
    * @returns {Promise<T>}
    * @memberof OperationQueue
    */
-  public async operate<T>(op: () => Promise<T>): Promise<T> {
+  public async operate<T>(
+    op: () => Promise<T>,
+    contextId?: string
+  ): Promise<T> {
     const id = uuid();
     this.operations.push({ id, op });
-    this.logger.debug(`OperationQueue.operate(${id})`);
+    this.logger.debug(
+      `OperationQueue.operate() Schedule incoming job ${id}`,
+      contextId
+    );
 
-    this.execute();
+    this.execute(contextId);
     return new Promise<T>((resolve, reject) => {
       this.emitter
         .once(id, res => {
           this.logger.debug(
-            `OperationQueue.operate(${id}) got callback, resolve`
+            `OperationQueue.operate() Job ${id} completes callback, resolve.`,
+            contextId
           );
 
           this.emitter.removeAllListeners("error_" + id);
           resolve(res);
 
           process.nextTick(() => {
-            this.execute();
+            this.execute(contextId);
           });
         })
         .once("error_" + id, err => {
-          this.logger.debug(`OperationQueue.operate(${id}) got error, reject`);
+          this.logger.debug(
+            `OperationQueue.operate() Job ${id} error, reject.`,
+            contextId
+          );
           this.emitter.removeAllListeners(id);
           reject(err);
           process.nextTick(() => {
-            this.execute();
+            this.execute(contextId);
           });
         });
     });
@@ -65,23 +76,21 @@ export default class OperationQueue implements IOperationQueue {
    * It assists queue to execute the operation.
    *
    * @private
+   * @param {string} [contextId]
    * @returns
    * @memberof OperationQueue
    */
-  private async execute<T>(): Promise<any> {
+  private async execute<T>(contextId?: string): Promise<any> {
     this.logger.debug(
-      `OperationQueue.execute() runningConcurrency:${
-        this.runningConcurrency
-      } maxConcurrency:${this.maxConcurrency}`
+      `OperationQueue:execute() Current runningConcurrency:${this.runningConcurrency} maxConcurrency:${this.maxConcurrency} operations.length:${this.operations.length}`,
+      contextId
     );
 
     if (this.runningConcurrency < this.maxConcurrency) {
-      this.logger.debug(
-        `OperationQueue.execute() operations.length:${this.operations.length}`
-      );
       if (this.operations.length === 0) {
         this.logger.debug(
-          `OperationQueue.execute() return. Operation.length === 0`
+          `OperationQueue:execute() return. Operation.length === 0`,
+          contextId
         );
         return;
       }
@@ -91,23 +100,23 @@ export default class OperationQueue implements IOperationQueue {
 
       let res;
       try {
-        this.logger.debug(`OperationQueue.execute() await ${head!.id}`);
+        // this.logger.debug(`OperationQueue:execute() await ${head!.id}`);
         res = await head!.op();
-        this.logger.debug(`OperationQueue.execute() ${head!.id} done`);
+        // this.logger.debug(`OperationQueue:execute() ${head!.id} done`);
       } catch (err) {
         this.runningConcurrency--;
-        this.logger.debug(
-          `OperationQueue.execute() ${head!.id} emit error ${err}`
-        );
+        // this.logger.debug(
+        //   `OperationQueue:execute() ${head!.id} emit error ${err}`
+        // );
         this.emitter.emit(`error_${head!.id}`, err);
         return;
       }
       this.runningConcurrency--;
-      this.logger.debug(
-        `OperationQueue.execute() emit ${head!.id}. runningConcurrency: ${
-          this.runningConcurrency
-        }`
-      );
+      // this.logger.debug(
+      //   `OperationQueue:execute() emit ${head!.id}. runningConcurrency: ${
+      //     this.runningConcurrency
+      //   }`
+      // );
       this.emitter.emit(head!.id, res);
     }
   }